fix(#221): extract command logic from cli/ into domain packages
Per eng/go §2.18 (INVARIANTS.md T-002): cli/ owns dispatch only.
Domain logic extracted to domain packages:

  cmd_doctor.go (238→44 lines) → internal/doctor/
  cmd_init.go   (131→27 lines) → internal/hubinit/
  cmd_status.go  (76→27 lines) → internal/hubstatus/
  cmd_deps.go    (82→67 lines) → FindIndexPath moved to internal/restore/

cmd_deps.go and cmd_build.go were already thin wrappers — verified
unchanged structurally (deps had FindIndexPath extracted).

CI dispatch boundary check added to .github/workflows/go.yml:
greps cmd_*.go for prohibited imports (os, net/http, encoding/json,
archive/, compress/, crypto/, path/filepath). Fails CI if any
cmd_*.go contains domain-level imports. Makes T-002 mechanical.

Each domain package has its own _test.go. Tests migrated from
cli/cmd_*_test.go to doctor/doctor_test.go, hubinit/hubinit_test.go,
hubstatus/hubstatus_test.go. cli/registry_test.go stays (tests
dispatch types).

Affected invariants:
- T-002 (kernel remains minimal): moving TOWARD — cli/ shrinks
- T-004 (source/artifact/installed explicit): no change

All tests pass. Dispatch boundary check clean.
go test ./... && go vet ./...
